Nested conditions/randomization + condition/rand persistence This commit makes two related changes to try to fix our timeline woes: (1) Randomization and condition assignments are persisted to local storage and synced to firestore. We do this to be able to access their values when `design.js` gets loaded, which is before we sync with firestore. Perhaps this is a slightly awkward design pattern, but to accomplish this, I keep parallel copies of this information in the `local` and `data` smilestore objects, sync them once when we `setDBConnected()` (see `manualSyncLocalToData`), and when I write to them, write to both if connected or to one if not (e.g., `setCondition`).
(2) Support nested randomization and condition assignments.
To do this, I introduce a second set of methods in `timeline.js` that are `registerX` instead of `pushX`, where `X \in {View, RandomizedNode, ConditionalNode}`.
The idea is to communicate that the two verbs have different semantics.
Registering a node means we let the timeline know about it without it necessarily being accessible or part of the view.
Pushing a node makes it part of the sequential timeline.
This allows to register randomized or conditional nodes, and then 'consume' them when pushing another node that refers to them.

We should test and document this more thoroughly once we agree that this is a reasonable way to do this, but it seems to work.