add blockIndex and blockLength
these getters give you a index within the current “block” of steps.  so if you have a heirarchically nested table and you are at a particularly leaf note blockIndex gives you the index of the parent node.  blocklength gives you the length of the parent node.  This is useful for stepping to the end of a block of trials without hardcoding how many trials are in each block.
- includes tests at the level of StepState and useViewAPI