resolving path/id confusion
path is the full path to a leaf /home/tree/trial
id is the name of a node (home, tree, trial)
when you append or create things you can add a id: field and if you do it will set the id of the node.
previously use used path for that which was confusing.